Remove error if seasonal == F | zonal ==F
Hello Badger! This is my first time forking, what fun.

When I ran the function without seasonal data, I found it produced an error due to looking for input$season_col from the first if statement. The same happens if zonal==F. It should work with this proposed change (and I wanted an excuse to finally learn what GitHub is all about in more detail). Cheers!